New components for some of the columns
