+++ src/pages/registration/files/FileTable.tsx
@@ -12,21 +12,13 @@ import {
 } from '@mui/material';
 import { useFormikContext } from 'formik';
 import { useTranslation } from 'react-i18next';
-import { useSelector } from 'react-redux';
-import { RootState } from '../../../redux/store';
-import { AssociatedFile, FileType, Uppy } from '../../../types/associatedArtifact.types';
+import { AssociatedFile, Uppy } from '../../../types/associatedArtifact.types';
 import { Registration } from '../../../types/registration.types';
 import { useFileTableColumnWidths } from '../../../utils/hooks/useFileTableColumnWidths';
-import {
-  isDegree,
-  isEmbargoed,
-  isSameFile,
-  isTypeWithFileVersionField,
-  isTypeWithRrs,
-  userCanUnpublishRegistration,
-} from '../../../utils/registration-helpers';
-import { FilesTableRow, markForPublishId } from './FilesTableRow';
+import { isSameFile, isTypeWithFileVersionField } from '../../../utils/registration-helpers';
+import { FileTableRow } from './FileTableRow';
 import { LicenseHelperModal } from './LicenseHelperModal';
+import { markForPublishId } from './PublishCheck';
 import { VersionHelperModal } from './VersionHelperModal';
 
 const StyledTableCell = styled(TableCell)({
@@ -52,36 +44,21 @@ interface FileListProps {
 export const FileTable = ({ files, uppy, remove, baseFieldName, archived }: FileListProps) => {
   const { t } = useTranslation();
   const { values, setFieldTouched } = useFormikContext<Registration>();
-  const { entityDescription, associatedArtifacts } = values;
 
-  const user = useSelector((store: RootState) => store.user);
-  const customer = useSelector((store: RootState) => store.customer);
-
-  const publicationInstanceType = entityDescription?.reference?.publicationInstance?.type;
-  const isProtectedDegree = isDegree(publicationInstanceType);
-  const showFileVersion = isTypeWithFileVersionField(publicationInstanceType);
-
+  const showFileVersion = isTypeWithFileVersionField(values.entityDescription?.reference?.publicationInstance?.type);
   const columnWidths = useFileTableColumnWidths(archived, showFileVersion);
 
-  function canEditFile(file: AssociatedFile) {
-    if (isProtectedDegree && isEmbargoed(file.embargoDate)) {
-      return !!user?.isEmbargoThesisCurator;
-    }
+  const removeFile = (associatedFileIndex: number, file: AssociatedFile) => {
+    const associatedArtifactsBeforeRemoval = values.associatedArtifacts.length;
+    const remainingFiles = uppy.getFiles().filter((uppyFile) => uppyFile.response?.uploadURL !== file.identifier);
+    uppy.setState({ files: remainingFiles });
+    remove(associatedFileIndex);
 
-    if (isProtectedDegree) {
-      return !!user?.isThesisCurator;
+    if (associatedArtifactsBeforeRemoval === 1) {
+      // Ensure field is set to touched even if it's empty
+      setFieldTouched(baseFieldName);
     }
-
-    if (values.type === 'ImportCandidate') {
-      return !!user?.isInternalImporter;
-    }
-
-    if (file.type === FileType.PublishedFile) {
-      return userCanUnpublishRegistration(values) ?? false;
-    }
-
-    return true;
-  }
+  };
 
   return (
     <TableContainer component={Paper} elevation={3} sx={{ mb: '2rem', width: '100%' }}>
@@ -118,29 +95,14 @@ export const FileTable = ({ files, uppy, remove, baseFieldName, archived }: File
         </TableHead>
         <TableBody>
           {files.map((file) => {
-            const associatedFileIndex = associatedArtifacts.findIndex((artifact) => isSameFile(artifact, file));
+            const associatedFileIndex = values.associatedArtifacts.findIndex((artifact) => isSameFile(artifact, file));
 
             return (
-              <FilesTableRow
+              <FileTableRow
                 key={file.identifier}
                 file={file}
-                disabled={!canEditFile(file)}
-                removeFile={() => {
-                  const associatedArtifactsBeforeRemoval = associatedArtifacts.length;
-                  const remainingFiles = uppy
-                    .getFiles()
-                    .filter((uppyFile) => uppyFile.response?.uploadURL !== file.identifier);
-                  uppy.setState({ files: remainingFiles });
-                  remove(associatedFileIndex);
-
-                  if (associatedArtifactsBeforeRemoval === 1) {
-                    // Ensure field is set to touched even if it's empty
-                    setFieldTouched(baseFieldName);
-                  }
-                }}
+                removeFile={() => removeFile(associatedFileIndex, file)}
                 baseFieldName={`${baseFieldName}[${associatedFileIndex}]`}
-                showFileVersion={showFileVersion}
-                showRrs={isTypeWithRrs(publicationInstanceType)}
                 archived={archived}
               />
             );
