+++ src/pages/registration/files/FileTableRow.tsx
@@ -1,14 +1,11 @@
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
-import InsertDriveFileOutlinedIcon from '@mui/icons-material/InsertDriveFileOutlined';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
 import {
   Box,
   Checkbox,
   Collapse,
-  FormControl,
   FormControlLabel,
-  FormHelperText,
   IconButton,
   Link as MuiLink,
   ListItemIcon,
@@ -16,8 +13,6 @@ import {
   MenuItem,
   Paper,
   Popover,
-  Radio,
-  RadioGroup,
   TableCell,
   TableCellProps,
   TableRow,
@@ -27,12 +22,9 @@ import {
 } from '@mui/material';
 import { DatePicker } from '@mui/x-date-pickers';
 import { ErrorMessage, Field, FieldProps, getIn, useFormikContext } from 'formik';
-import prettyBytes from 'pretty-bytes';
 import { useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
-import { ConfirmDialog } from '../../../components/ConfirmDialog';
-import { TruncatableTypography } from '../../../components/TruncatableTypography';
 import { RootState } from '../../../redux/store';
 import { AssociatedFile, FileRrs, FileType, FileVersion } from '../../../types/associatedArtifact.types';
 import { CustomerRrsType } from '../../../types/customerInstitution.types';
@@ -42,8 +34,16 @@ import { Registration } from '../../../types/registration.types';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { equalUris } from '../../../utils/general-helpers';
 import { useFileTableColumnWidths } from '../../../utils/hooks/useFileTableColumnWidths';
-import { DeleteIconButton } from '../../messages/components/DeleteIconButton';
-import { DownloadFileButton } from '../files_and_license_tab/DownloadFileButton';
+import {
+  isDegree,
+  isEmbargoed,
+  isTypeWithFileVersionField,
+  isTypeWithRrs,
+  userCanUnpublishRegistration,
+} from '../../../utils/registration-helpers';
+import { FileInfo } from './FileInfo';
+import { PublishCheck } from './PublishCheck';
+import { VersionRadio } from './VersionRadio';
 
 export const markForPublishId = 'mark-for-publish';
 
@@ -55,33 +55,20 @@ interface FilesTableRowProps {
   file: AssociatedFile;
   removeFile: () => void;
   baseFieldName: string;
-  showFileVersion: boolean;
-  showRrs: boolean;
-  disabled: boolean;
   archived?: boolean;
 }
 
-export const FilesTableRow = ({
-  file,
-  removeFile,
-  baseFieldName,
-  showFileVersion,
-  showRrs,
-  disabled,
-  archived,
-}: FilesTableRowProps) => {
+export const FileTableRow = ({ file, removeFile, baseFieldName, archived }: FilesTableRowProps) => {
   const { t } = useTranslation();
+  const { values } = useFormikContext<Registration>();
   const user = useSelector((state: RootState) => state.user);
   const customer = useSelector((state: RootState) => state.customer);
-  const [openConfirmDialog, setOpenConfirmDialog] = useState(false);
-  const toggleOpenConfirmDialog = () => setOpenConfirmDialog(!openConfirmDialog);
+
   const { setFieldValue, setFieldTouched, errors, touched } = useFormikContext<Registration>();
   const columnWidths = useFileTableColumnWidths(archived);
 
   console.log('columnWidths', columnWidths);
 
-  const fileTypeFieldName = `${baseFieldName}.${SpecificFileFieldNames.Type}`;
-  const publisherVersionFieldName = `${baseFieldName}.${SpecificFileFieldNames.PublisherVersion}`;
   const licenseFieldName = `${baseFieldName}.${SpecificFileFieldNames.License}`;
   const embargoFieldName = `${baseFieldName}.${SpecificFileFieldNames.EmbargoDate}`;
   const legalNoteFieldName = `${baseFieldName}.${SpecificFileFieldNames.LegalNote}`;
@@ -114,125 +101,51 @@ export const FilesTableRow = ({
   );
   const inactiveLicenses = licenses.filter((license) => license.version && license.version !== 4);
 
+  const publicationInstanceType = values.entityDescription?.reference?.publicationInstance?.type;
+  const isProtectedDegree = isDegree(publicationInstanceType);
+
+  function canEditFile(file: AssociatedFile) {
+    if (isProtectedDegree && isEmbargoed(file.embargoDate)) {
+      return !!user?.isEmbargoThesisCurator;
+    }
+
+    if (isProtectedDegree) {
+      return !!user?.isThesisCurator;
+    }
+
+    if (values.type === 'ImportCandidate') {
+      return !!user?.isInternalImporter;
+    }
+
+    if (file.type === FileType.PublishedFile) {
+      return userCanUnpublishRegistration(values) ?? false;
+    }
+
+    return true;
+  }
+
+  const disabled = !canEditFile(file);
+
+  const showFileVersion = isTypeWithFileVersionField(publicationInstanceType);
+  const showRrs = isTypeWithRrs(publicationInstanceType);
+
   return (
     <>
       <TableRow
         data-testid={dataTestId.registrationWizard.files.fileRow}
         title={disabled ? t('registration.files_and_license.disabled_helper_text') : ''}
-        sx={{
-          bgcolor: disabled ? 'grey.400' : '',
-          td: { verticalAlign: 'top', borderBottom: !archived ? 'unset' : '' },
-        }}>
+        sx={{ bgcolor: disabled ? 'grey.400' : '' }}>
         <VerticalAlignedTableCell sx={{ minWidth: columnWidths.nameColumn + '%' }}>
-          <Box sx={{ display: 'flex', minWidth: '13rem', gap: '0.75rem', alignItems: 'center' }}>
-            <InsertDriveFileOutlinedIcon sx={{ color: disabled ? 'grey.600' : '' }} />
-            <Box sx={{ minWidth: '10rem' }}>
-              <TruncatableTypography sx={{ fontWeight: 'bold', color: disabled ? 'grey.600' : '' }}>
-                {file.name}
-              </TruncatableTypography>
-              <Typography sx={{ color: disabled ? 'grey.600' : '' }}>{prettyBytes(file.size)}</Typography>
-            </Box>
-            <Box sx={{ minWidth: '1.5rem' }}>
-              <DownloadFileButton file={file} greyTones={disabled} />
-            </Box>
-            <DeleteIconButton
-              data-testid={dataTestId.registrationWizard.files.deleteFile}
-              onClick={disabled ? undefined : toggleOpenConfirmDialog}
-              tooltip={t('registration.files_and_license.remove_file')}
-              disabled={disabled}
-            />
-            <ConfirmDialog
-              open={openConfirmDialog}
-              title={t('registration.files_and_license.remove_file')}
-              onAccept={() => {
-                removeFile();
-                toggleOpenConfirmDialog();
-              }}
-              onCancel={toggleOpenConfirmDialog}>
-              <Typography>
-                {t('registration.files_and_license.remove_file_description', { fileName: file.name })}
-              </Typography>
-            </ConfirmDialog>
-          </Box>
+          <FileInfo disabled={disabled} file={file} removeFile={removeFile} />
         </VerticalAlignedTableCell>
         <VerticalAlignedTableCell sx={{ width: columnWidths.publishColumn + '%' }}>
-          <Field name={fileTypeFieldName}>
-            {({ field }: FieldProps) => (
-              <Checkbox
-                {...field}
-                data-testid={dataTestId.registrationWizard.files.toPublishCheckbox}
-                checked={field.value === FileType.UnpublishedFile || field.value === FileType.PublishedFile}
-                disabled={disabled}
-                inputProps={{
-                  'aria-labelledby': markForPublishId,
-                }}
-                onChange={(_, checked) => {
-                  if (!checked) {
-                    setFieldValue(fileTypeFieldName, FileType.UnpublishableFile);
-                  } else {
-                    setFieldValue(fileTypeFieldName, FileType.UnpublishedFile);
-                  }
-                }}
-              />
-            )}
-          </Field>
+          <PublishCheck disabled={disabled} baseFieldName={baseFieldName} />
         </VerticalAlignedTableCell>
         {!archived && (
           <>
             {showFileVersion && (
               <VerticalAlignedTableCell sx={{ width: columnWidths.versionColumn + '%' }}>
-                <Field name={publisherVersionFieldName}>
-                  {({ field, meta: { error, touched } }: FieldProps<FileVersion | null>) => (
-                    <FormControl data-testid={dataTestId.registrationWizard.files.version} required disabled={disabled}>
-                      <RadioGroup
-                        {...field}
-                        row
-                        sx={{ flexWrap: 'nowrap', flexDirection: { lg: 'row', md: 'column' } }}
-                        onChange={(event) => {
-                          const fileVersion = event.target.value as FileVersion;
-                          setFieldValue(field.name, fileVersion);
-
-                          if (showRrs) {
-                            if (fileVersion === FileVersion.Published) {
-                              const nullRrsValue: FileRrs = {
-                                type: 'NullRightsRetentionStrategy',
-                                configuredType: rrsStrategy,
-                              };
-                              setFieldValue(rrsFieldName, nullRrsValue);
-                              setFieldValue(licenseFieldName, null);
-                            } else if (isCustomerRrs || isOverridableRrs) {
-                              const customerRrsValue: FileRrs = {
-                                type: 'CustomerRightsRetentionStrategy',
-                                configuredType: rrsStrategy,
-                              };
-                              setFieldValue(rrsFieldName, customerRrsValue);
-                              setFieldValue(licenseFieldName, LicenseUri.CC_BY_4);
-                            }
-                          }
-                        }}>
-                        <FormControlLabel
-                          value={FileVersion.Accepted}
-                          control={<Radio />}
-                          label={
-                            <Typography variant="caption" sx={{ lineHeight: '1.1rem' }}>
-                              {t('registration.files_and_license.accepted_version')}
-                            </Typography>
-                          }
-                        />
-                        <FormControlLabel
-                          value={FileVersion.Published}
-                          control={<Radio />}
-                          label={
-                            <Typography variant="caption" sx={{ lineHeight: '1.1rem' }}>
-                              {t('registration.files_and_license.published_version')}
-                            </Typography>
-                          }
-                        />
-                      </RadioGroup>
-                      {error && touched && <FormHelperText error>{error}</FormHelperText>}
-                    </FormControl>
-                  )}
-                </Field>
+                <VersionRadio disabled={disabled} baseFieldName={baseFieldName} rrsStrategy={rrsStrategy} />
               </VerticalAlignedTableCell>
             )}
             <VerticalAlignedTableCell sx={{ width: columnWidths.licenseColumn + '%' }}>
